backend/requirements.txt
Django==4.2
djangorestframework==3.15.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.4.0
python-dotenv==1.0.0
dj-database-url==2.1.0
mysqlclient==2.2.0
setuptools==68.0.0